system: Move logging/MQTT publish for ControlsLock to Cmd.cpp
This is more a cleanup than a fix. The only thing it fixes that
System_IndicateOk is now executed when enabling and disabling Controls
Lock.

However looking at the implementation of other modificators or MQTT
commands it's more consistent if the logging of the modificator and
sending the MQTT message happens where the commands are triggered
(Cmd.cpp and Mqtt.cpp).