revision: Add Git revision to logs
This change adds a pre build script that generates a gitrevision.h
header with the current git revision. That is then logged in the serial
console and web log.

When comparing log files (for instance to hunt issues) it's useful to
know exactly the git version the software has been compiled with. If
local modifications have been done, '-dirty' is added to the version
so that uncommitted changes can easily be identified.