mqtt: Add topicCoverChangedState
Publishes an (empty) message to topciCoverChangedState whenever the
cover image has changed. This allows the MQTT client application to
know when it should reload the cover image.

Note that due to performance reasons, the MQTT client application is
responsible to only load the cover image if it is currently visible to
the user.