Led.cpp: Refactor volume level indication and fix small glitch
This refactors the code that shows the volume level on the LEDs.
Instead of waiting for further volume changes and then showing
them in the next interation of the outer task loop, we simply
loop and redraw in the tight loop.

This also fixes a glitch where the LEDs briefly show the default
pattern during volume control.