Add initial Kconfig support for compile-time configuration
Instead of modifying the preprocessor macros in the settings header
files manually, this introduces support for Kconfig which provides
a common and widely used mechanism to manage the compile-time
settings.

To support this the Kconfiglib Python library is installed in the
PlatformIO virtual env using a script.

To change the configuration via the interactive menu, run the
following command in a PlatformIO terminal:

    pio run -t menuconfig