Kconfig: Add hack to translate string literals to identifiers
Kconfig can only deal with literals, not with references to preprocessor
macros. For some purposes we need to set/select a predefined macro. To achieve
this we use a string literal in Kconfig and then translate the value (remove
string quotes) here. The translated value is replaced before generating the
config.h header.

To enable the automatic translation, make the config item depend on the
CONVERT_TO_TOKEN item in the Kconfig file.