Bluetooth.cpp: Fix logging crash in sink mode
In sink mode a2dp_source is NULL which causes a crash when the state
change callbacks are called.

To fix this pass the correct pointer to the sink or source object
via the callback state variable and use a cast to BluetoothA2DPCommon
for accessing the to_str() method.