stricter hostname validation
Changes the frontend validation and adds a backend validation.
The backend previously did not validate the hostname at all.

The hostname validation is primarily for mDNS. For the "pretty" hostname
much more could be allowed, but we do not make that distinction (yet).

In general, change what kind of hostnames are accepted.
For example, a hostname can have a '-' in the middle.
Also, only alphanumerical characters are allowed. Dots ('.') are still
disallowed to prevent DNS problems when using mDNS.