disable PORT_ExpanderISR until the input registers have been read
The interrupt line is active low so switch from FALLING to ONLOW.
To then prevent continuous calling of the ISR, it needs to be disabled.
Port_ExpanderHandler() re-enables it after handling things. It also
takes care to enable it in the first place.

While at it, use digitalPinToInterrupt() to convert the pin to the
interrupt number as recommended by the Arduino reference manual
(https://www.arduino.cc/reference/en/language/functions/external-interrupts/attachinterrupt/).