Rework playlist generation (#275)
* Rename and adapt playlist destruction function

Rename freeMultiCharArray to freePlaylist and specialise it so that it can be used to destroy the content of a playlist.
Move all playlist related functions into it's own header

* Optimise playlist generation

Use std::vector instead of POD char* array so that we do not have to care about memory managment.
We still have to take care of the strings in the playlist manually.

Make x_malloc more forgiving by prefereing the PSRAM and not enforcing it.

* Change playlist to std::vector pointer

* rework SdCard_pickRandomSubdirectory

Rework SdCard_pickRandomSubdirectory to fir into the new playlist generator.

* Rework TrackQueueDispatcher

Make the TrackQueueDispatcher aware of the new std::vector playlist.
Adapt FreeRTOS queue to use Playlist pointers.
Improve log output of sorting function.

* Make the AudioPlayer Task aware of the std::vector

Fix compile errors when accessing playlist std::vector. Rework the shuffle and sort function to use the std-lib.

* Make cover handler aware of the std::vector

* Remove numberOfTracks and use vector size instead

Remove the numberOfTracks, since it's redundant with std::vector->size()

* Rework and improve fileValid

* Alway supply an empty Playlist

Make sure there is always a dummy playlist and move initialization of gPlayProperties into the AudioPlayer Task.
Fix formatting issues and the review points

- Add guards to prevent access to playlist if pointer is nullptr. This will be improved in PR 2 of 3 by encapsulating playlist and preventing direct access.
- remove commented code parts
- fix spelling

* Reduce logging of invalid files

Comment the logging of invalid files to reduce the log spam. Add logging of the number of ignored files in the playlist generation.