Add coverimage support for flac files in Web-UI
This adds cover support for "normal" flac files in Web-UI. In this "normal" case the cover is stored in PICTURE block-type. The cover is not encoded, similar to ID3 Tag in mp3.

Since the cover header in flac files has another structure than the cover header in mp3 files, mp3 files and flac files have to be distinguished.
mp3s are identified by the first three bytes in header ("ID3"), which is only valid für ID3v2 mp3s. This is ok since all mp3s with covers are ID3v2 (and at this point in code only files with covers are treated).

flacs are identified by the first four bytes in header ("fLaC").

For the cover header in flac files see e.g. https://datatracker.ietf.org/doc/draft-ietf-cellar-flac/, page 28. The header consists of a "metadata block header", specifying the cover filesize including header (3 Bytes), followed by the elements of the table.

There may also be an "abnormal" case, where the cover is stored in a VORBIS COMMENT block with a base-64 encoded METADATA_BLOCK_PICTURE tag (the way pictures are embedded in Ogg/Vorbis). This can be treated by another PR.