Fix playing a webstream while still connecting to WiFi
The code is simplified a lot. I removed the functions and variables that
were never running anyway.

The webstream command is now always accepted, and the error is handled
inside the audio task. The audio task also blocks as long as a
connection try is in progress.

In addition to the code that did not work anyway, this should also fix
the oversight that we want to immediately play a different webstream
after starting, not only the last RFID.