Update settings.h
Adjusted MQTT settings for easier integration of multiple ESPuino devices in a household.
The default setting #define DEVICE_HOSTNAME "ESP32-ESPuino" is now used as the MQTT-specific device name. This hostname is also utilized in the MQTT topics. As a result, integration with Home Assistant is now simpler and more streamlined.